TFE_Jedi: fix 32bit build
On 32bit, the size_t min/max macros conflict with the u32 ones
(since size_t is u32 on 32bit).
Change the type from 'size_t' to 'u64' which is the correct type
on 64bit.
Fixes build for 32bit platforms.

+++ TheForceEngine/TFE_Jedi/Math/core_math.h
@@ -338,17 +338,17 @@ namespace TFE_Jedi
 	}
 
 	// Size_t
-	inline size_t min(size_t a, size_t b)
+	inline u64 min(u64 a, u64 b)
 	{
 		return a < b ? a : b;
 	}
 
-	inline size_t max(size_t a, size_t b)
+	inline u64 max(u64 a, u64 b)
 	{
 		return a > b ? a : b;
 	}
 
-	inline size_t clamp(size_t x, size_t a, size_t b)
+	inline u64 clamp(u64 x, u64 a, u64 b)
 	{
 		return min(max(x, a), b);
 	}
